src/upload_api/requirements.txt
fastapi[all]==0.103.1
pydantic-settings

zipfile38==0.0.3

azure-storage-blob==12.8.0
azure-identity==1.14.0